Split particle and fluid solvers into tests/particles, rename the backend module, and relocate the gravity setter test.
