+++ CODING_GUIDELINES.md
@@ -169,7 +169,7 @@ Genesis strongly resists the accumulation of small wrappers. Helpers accrete, ge
 
 ### 12.2 Test structure
 
-- **Tests are organized per component, then per capability.** Each component has a folder under `tests/` (`rigid/`, `deformable/`, `ipc/`, `coupling/`, `sensors/`, `rendering/`, `loaders/`, `core/`, `integration/`, `benchmarks/`) holding one file per capability (for example `tests/rigid/test_collision.py`, `tests/sensors/test_imu.py`); shared model fixtures live in the folder's `conftest.py`. New tests go into the existing capability file that covers them; a new file is only warranted for a genuinely new capability, never a fresh `test_<one_thing>.py` for a single feature.
+- **Tests are organized per component, then per capability.** Each component has a folder under `tests/` (`rigid/`, `deformable/`, `particles/`, `ipc/`, `coupling/`, `sensors/`, `rendering/`, `loaders/`, `core/`, `integration/`, `benchmarks/`) holding one file per capability (for example `tests/rigid/test_collision.py`, `tests/sensors/test_imu.py`); shared model fixtures live in the folder's `conftest.py`. New tests go into the existing capability file that covers them; a new file is only warranted for a genuinely new capability, never a fresh `test_<one_thing>.py` for a single feature.
 - **Pack tests.** Prefer one comprehensive scene with diverse entities and options over many small single-option tests; add differently-configured entities to the same scene rather than writing sibling functions. One scene build per test: each additional build recompiles kernels (roughly 18 s locally, minutes on CI), and folding reference simulations into the same scene makes comparisons tighter, not looser.
 - **Prefer parametrization over new test functions.** A new case for the same capability becomes a `@pytest.mark.parametrize` dimension on the existing test. Parametrize batched tests over `n_envs=[0, 2]` - multi-env is where shape bugs hide - but do not add a dead dimension that a conftest fixture already controls (such as `backend`).
 - **Name tests after the feature validated, not the scenario used to validate it.** A test exercising off-axis contact rejection through a stacking-tower scene is `test_reject_offaxis_contact_on_authored_decomp`, not `test_stacking_tower_stability`; the tower is *how*, the rejection is *what*. Never repeat the module or folder name as a prefix: `tests/sensors/test_temperature.py::test_grid_sensor_contact_and_reset`, never `test_temperature_grid_...`.
